+++ pkgs/ffigen/test/native_objc_test/static_func_test.m
@@ -58,3 +58,6 @@ void staticFuncConsumesArg(StaticFuncTestObj* __attribute((ns_consumed)) a) {
   if (counter != nil) --*counter;
 }
 @end
+
+int foo(int x) { return 10 * x; }
+int fooPtr(int x) { return 100 * x; }