+++ pkgs/jni/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.15.2
+
+- Do not fail `flutter build` if JDK is not found for desktop.
+
+## 0.15.1
+
+- **Breaking Change**: Removed the `engineId` argument from
+  `Jni.androidApplicationContext` as the application context is not dependant on
+  the engine ID.
+
 ## 0.15.0
 
 - **Breaking Change**: Made `Jni.env` internal.